added laplace and vc-laplace scripts

+++ c++/rchol_lap/rchol_lap.cpp
@@ -9,8 +9,8 @@
 #include "spvec.hpp"
 
 
-typedef float real;
-//typedef double real;
+//typedef float real;
+typedef double real;
 typedef spvec<int, real> edges;
